% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedivieweR_app.R
\name{pedivieweR_app}
\alias{pedivieweR_app}
\title{Create Pedigree Viewer Shiny Application}
\usage{
pedivieweR_app()
}
\value{
A Shiny application object
}
\description{
This function creates and returns a Shiny application object for
the Pedigree Viewer. The app can be run using \code{\link{run_pedivieweR}}
or by passing the returned object to \code{\link[shiny]{runApp}}.
}
